split off work.py
